+++ guillotina_kafka/subscribers.py
@@ -1,7 +1,9 @@
-from guillotina.interfaces import IApplicationCleanupEvent
-from guillotina.component import get_utility
 from guillotina import configure
+from guillotina.component import get_utility
+from guillotina.interfaces import IApplicationCleanupEvent
+
 from guillotina_kafka import IKafkaProducerUtility
+
 
 @configure.subscriber(for_=(IApplicationCleanupEvent))
 async def app_cleanup(event):