+++ guillotina_kafka/tests/conftest.py
@@ -1,5 +1,5 @@
 pytest_plugins = [
-    'guillotina.tests.fixtures',
-    'guillotina_kafka.tests.fixtures',
-    'pytest_docker_fixtures',
+    "guillotina.tests.fixtures",
+    "guillotina_kafka.tests.fixtures",
+    "pytest_docker_fixtures",
 ]