+++ guillotina_kafka/tests/test_consumer.py
@@ -1,21 +1,22 @@
+import pytest
+from aiokafka.structs import ConsumerRecord, RecordMetadata
 from guillotina import app_settings
-from aiokafka.structs import RecordMetadata
-from aiokafka.structs import ConsumerRecord
 from guillotina.component import get_adapter
-from guillotina_kafka.producer import GetKafkaProducer
-from guillotina_kafka.interfaces import IConsumerUtility
+
 from guillotina_kafka.consumer.stream import StreamConsumer
+from guillotina_kafka.interfaces import IConsumerUtility
+from guillotina_kafka.producer import GetKafkaProducer
 
-
+pytestmark = pytest.mark.asyncio
 
 
 async def test_stream_consumer(kafka_container, event_loop, container_requester):
 
-    TEST_TOPICS = ['test-topic-1', 'test-topic-2']
-    TEST_GROUP = 'test-group'
-    BOOTSTRAP_SERVERS=app_settings['kafka']['brokers']
+    TEST_TOPICS = ["test-topic-1", "test-topic-2"]
+    TEST_GROUP = "test-group"
+    BOOTSTRAP_SERVERS = app_settings["kafka"]["brokers"]
 
-    producer = GetKafkaProducer('bytes', app_settings)
+    producer = GetKafkaProducer("bytes", app_settings)
 
     async def worker_func(record):
         assert isinstance(record, ConsumerRecord)
@@ -25,12 +26,12 @@ async def test_stream_consumer(kafka_container, event_loop, container_requester)
     RECORDS_SENT = {}
 
     for index, topic in enumerate(TEST_TOPICS):
-        result = await producer.send(topic, value=f'foobar{index}')
+        result = await producer.send(topic, value=f"foobar{index}")
         assert isinstance(result, tuple)
         assert result[0] is True
         assert isinstance(result[1], RecordMetadata)
         assert result[1].topic == topic
-        RECORDS_SENT.setdefault(result[1].topic, []).append(f'foobar{index}'.encode())
+        RECORDS_SENT.setdefault(result[1].topic, []).append(f"foobar{index}".encode())
 
     await producer.stop()
 
@@ -38,10 +39,10 @@ async def test_stream_consumer(kafka_container, event_loop, container_requester)
         TEST_TOPICS,
         worker=worker_func,
         group_id=TEST_GROUP,
-        bootstrap_servers=BOOTSTRAP_SERVERS
+        bootstrap_servers=BOOTSTRAP_SERVERS,
     )
 
-    consumer = get_adapter(consumer, IConsumerUtility, name='stream')
+    consumer = get_adapter(consumer, IConsumerUtility, name="stream")
     assert consumer.consumer.is_ready is False
     await consumer.consumer.init()
     assert consumer.consumer.is_ready is True