+++ guillotina_kafka/utilities.py
@@ -1,14 +1,12 @@
-from aiokafka import AIOKafkaProducer
-from guillotina import app_settings
-from guillotina import configure
-from guillotina.component import get_utilities_for
-from guillotina.component import get_utility
-from guillotina_kafka.interfaces import IActiveConsumer
-from guillotina_kafka.interfaces import IKafkaProducerUtility
-
 import asyncio
 import json
 import logging
+
+from aiokafka import AIOKafkaProducer
+from guillotina import app_settings, configure
+from guillotina.component import get_utilities_for, get_utility
+
+from guillotina_kafka.interfaces import IActiveConsumer, IKafkaProducerUtility
 
 logger = logging.getLogger(__name__)
 
@@ -18,6 +16,7 @@ class KafkaProducerUtility:
     """This defines the singleton that will hold the connection to kafka
     and allows to send messages from it.
     """
+
     def __init__(self, loop=None):
         # Get kafka connection details from app settings
         self.loop = loop
@@ -26,9 +25,11 @@ class KafkaProducerUtility:
 
     async def stop_active_consumers(self):
         for name, worker in get_utilities_for(IActiveConsumer):
-            if hasattr(worker, '__consumer__') and not getattr(worker, '__stopped__', False):
+            if hasattr(worker, "__consumer__") and not getattr(
+                worker, "__stopped__", False
+            ):
                 try:
-                    logging.warning(f'Stopping {name} consumer')
+                    logging.warning(f"Stopping {name} consumer")
                     await worker.__consumer__.stop()
                     worker.__stopped__ = True
                 except Exception:
@@ -46,12 +47,14 @@ class KafkaProducerUtility:
             # make configuration is locked so multiple tasks can't attempt
             if self.is_ready:
                 return
-            self.config = {**{
-                'bootstrap_servers': app_settings['kafka']['brokers'],
-                'value_serializer': lambda data: json.dumps(data).encode('utf-8')
-            }, **kwargs}
-            self.config.setdefault(
-                'loop', self.loop or asyncio.get_event_loop())
+            self.config = {
+                **{
+                    "bootstrap_servers": app_settings["kafka"]["brokers"],
+                    "value_serializer": lambda data: json.dumps(data).encode("utf-8"),
+                },
+                **kwargs,
+            }
+            self.config.setdefault("loop", self.loop or asyncio.get_event_loop())
             if self.producer is None:
                 producer = AIOKafkaProducer(**self.config)
                 await producer.start()
@@ -70,12 +73,12 @@ class KafkaProducerUtility:
 
     async def send(self, *args, **kwargs):
         if not self.is_ready:
-            raise Exception('Producer utility is not ready')
+            raise Exception("Producer utility is not ready")
         return await self.producer.send(*args, **kwargs)
 
     async def stop(self):
         if not self.is_ready:
-            raise Exception('Producer utility is not ready')
+            raise Exception("Producer utility is not ready")
         _ = await self.producer.stop()
         self.producer = None
         return _